requirements.txt
# AI Assistant Web App - Dependencies
# Python 3.10+ recommended

# Web framework
flask>=3.0.0

# LangChain and LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Environment and utilities
python-dotenv>=1.0.0
